requirements.txt
# Core dependencies
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Azure AI
azure-ai-inference>=1.0.0
azure-core>=1.29.0

# Sentence transformers for STS
sentence-transformers>=2.2.0
torch>=2.0.0

# Utilities
numpy>=1.24.0
